bugs fixed

+++ api/src/main/java/br/com/med/voll/api/repository/ConsultationRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.time.LocalDateTime;
 
-public interface ConsultationRepository  extends JpaRepository<Consultation, Integer> {
+public interface ConsultationRepository  extends JpaRepository<Consultation, Long> {
 
     boolean existsByDoctorIdAndDate(Long doctorId, LocalDateTime date);
